mebot/requirements.txt
blis==0.4.1
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
Click==7.0
colorama==0.4.3
colorful==0.5.4
cymem==2.0.3
en-core-web-sm==2.3.0
Flask==1.1.2
Flask-Cors==3.0.8
idna==2.9
importlib-metadata==1.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.15.1
MarkupSafe==1.1.1
murmurhash==1.0.2
neobolt==1.7.17
neotime==1.7.4
numpy==1.19.0
pandas==1.0.5
plac==1.1.3
preshed==3.0.2
prompt-toolkit==2.0.10
py2neo==4.3.0
Pygments==2.3.1
python-aiml==0.9.3
python-dateutil==2.8.1
pytz==2020.1
requests==2.24.0
scikit-learn==0.23.1
scipy==1.4.1
six==1.15.0
sklearn==0.0
spacy==2.3.0
srsly==1.0.2
thinc==7.4.1
threadpoolctl==2.1.0
tqdm==4.46.1
urllib3==1.24.3
wasabi==0.6.0
wcwidth==0.2.4
Werkzeug==1.0.1
zipp==3.1.0
